+++ internal/report/sarif.go
@@ -143,7 +143,7 @@ func GenerateSarif(r *Report) *SarifLog {
 				Tool: SarifTool{
 					Driver: SarifDriver{
 						Name:           "VibeGuard",
-						Version:        "4.4.0",
+						Version:        "4.5.0",
 						InformationURI: "https://github.com/Ranaveer9177/cyberhackathon",
 						Rules:          rules,
 					},